create a travis.Build class
