+++ grading/pull_request.py
@@ -1,8 +1,6 @@
 import functools
 import github
-import json
-import re
-import requests
+import travis
 
 
 class PullRequest:
@@ -39,19 +37,6 @@ class PullRequest:
         return int(self.travis_url().split('/')[-1])
 
     @functools.lru_cache()
-    def travis_build_data(self):
-        url = "https://api.travis-ci.org/repos/{}/builds?ids={:d}".format(self.base_repo(), self.travis_build_id())
-        headers = {'User-Agent': 'MyClient/1.0.0', 'Accept': 'application/vnd.travis-ci.2+json'}
-        response = requests.get(url, headers=headers)
-        data = response.text
-        return json.loads(data)
-
-    @functools.lru_cache()
-    def travis_job_id(self):
-        builds = self.travis_build_data()['builds']
-        return int(builds[0]['job_ids'][0])
-
-    @functools.lru_cache()
     def modified_files_data(self):
         url = "https://api.github.com/repos/{}/pulls/{:d}/files".format(self.base_repo(), self.data['number'])
         return github.api_request(url)
@@ -62,29 +47,7 @@ class PullRequest:
         return map(lambda x: x['filename'], files)
 
     @functools.lru_cache()
-    def travis_log_url(self):
-        job_id = self.travis_job_id()
-        if isinstance(job_id, int) is False:
-            pytest_report = {}
-            print("wrong S3 id for user %s." % self.user())
-            return pytest_report
-        return "https://s3.amazonaws.com/archive.travis-ci.org/jobs/{:d}/log.txt".format(job_id)
-
-    @functools.lru_cache()
-    def travis_log(self):
-        """Retrieves the raw log data from S3."""
-        url = self.travis_log_url()
-        response = requests.get(url)
-        return response.text
-
-    @functools.lru_cache()
-    def pytest_report(self):
-        log_data = self.travis_log().replace('\n', ' ').replace('\r', ' ')
-        pattern = re.compile('<MQkrXV>[^{]+([{].*[}]{2,3}) ', re.MULTILINE)
-        match = pattern.search(log_data)
-        # TODO ensure it's only found once
-        if match:
-            report = match.group(1)
-            return json.loads(report)
-        else:
-            return None
+    def travis_build(self):
+        repo = self.base_repo()
+        build_id = self.travis_build_id()
+        return travis.Build(repo, build_id)
